requirements.txt
aiohttp==3.9.1
discord.py==2.3.2
feedparser==6.0.11
fluent.runtime==0.4.0